workflows/eval: Make sure to compare against the push run
For PRs that get merged without a merge commit like
https://github.com/NixOS/nixpkgs/commit/bcc5c141bf43460909b83740e96ceeea6d0f7e12,
the eval workflow has two associated runs for the commit, one for the
push event and one for the pull request event.

The pull request event doesn't run fully because the PR is already closed,
the push event is the one that will run fully and should be used to compare against.
So far it didn't distinguish between the two events, causing it to
sometimes pick the wrong one and then fail to download non-existent
artifacts: https://github.com/NixOS/nixpkgs/actions/runs/12092053414/job/33721377179

This commit fixes that by ensuring that the push event result is
compared against