treewide: format all inactive Nix files
Inactive means that there are no open PRs with activity in the
last month that touch those files.
A bunch later, we can do another pass to get the rest.

Doing it this way makes ensures that we don't cause any conflicts for
recently active PRs that would be ready to merge.
Only once those PRs get updated, CI will kick in and require the files
to be formatted.

Furthermore, this makes sure that we can merge this PR without having
to constantly rebase it!

This can be verified using

    nix-build https://github.com/infinisil/treewide-nixpkgs-reformat-script/archive/$Format:%H$.tar.gz \
      --argstr baseRev 314edfd2d5dcca142d4e1bec52e96623120ee9ab
    result/bin/apply-formatting $NIXPKGS_PATH